requirements.txt
requests
websocket-client
zstandard
brotli
